+++ comments.php
@@ -17,7 +17,7 @@ if ( post_password_required() ) {
 	return;
 }
 ?>
-
+<center>
 <div id="comments" class="comments-area">
 
 	<?php // You can start editing here -- including this comment! ?>
@@ -67,3 +67,4 @@ if ( post_password_required() ) {
 	<?php comment_form(); ?>
 
 </div><!-- #comments -->
+</center>