+++ functions.php
@@ -85,7 +85,7 @@ add_action( 'after_setup_theme', 'httpdjmcloud_danieljmckeown_com_setup' );
  */
 function httpdjmcloud_danieljmckeown_com_widgets_init() {
 	register_sidebar( array(
-		'name'          => __( 'Sidebar', 'httpdjmcloud-danieljmckeown-com' ),
+		'name'          => __( 'Below Content Left Sidebar', 'httpdjmcloud-danieljmckeown-com' ),
 		'id'            => 'sidebar-1',
 		'description'   => '',
 		'before_widget' => '<aside id="%1$s" class="widget %2$s">',
@@ -94,8 +94,17 @@ function httpdjmcloud_danieljmckeown_com_widgets_init() {
 		'after_title'   => '</h1>',
 	) );
 	register_sidebar( array(
-		'name'          => __( 'Sidebar2', 'httpdjmcloud-danieljmckeown-com' ),
+		'name'          => __( 'Below Content Right Sidebar', 'httpdjmcloud-danieljmckeown-com' ),
 		'id'            => 'sidebar-2',
+		'description'   => '',
+		'before_widget' => '<aside id="%1$s" class="widget %2$s">',
+		'after_widget'  => '</aside>',
+		'before_title'  => '<h1 class="widget-title">',
+		'after_title'   => '</h1>',
+	) );
+	register_sidebar( array(
+		'name'          => __( 'Responsive Left Sidebar', 'httpdjmcloud-danieljmckeown-com' ),
+		'id'            => 'sidebar-3',
 		'description'   => '',
 		'before_widget' => '<aside id="%1$s" class="widget %2$s">',
 		'after_widget'  => '</aside>',