+++ header.php
@@ -17,6 +17,7 @@
 
     <link href="<?php bloginfo('template_directory'); ?>/css/materialize.css" type="text/css" rel="stylesheet" media="screen,projection"/>
     <link href="<?php bloginfo('template_directory'); ?>/css/style.css" type="text/css" rel="stylesheet" media="screen,projection"/>
+    <link href="<?php bloginfo('template_directory'); ?>/css/portagebay.css" type="text/css" rel="stylesheet" media="screen,projection"/>
 
 </head>
 
@@ -27,19 +28,17 @@
 	<a class="skip-link screen-reader-text" href="#content"><?php _e( 'Skip to content', 'httpdjmcloud-danieljmckeown-com' ); ?></a>    
 
 	<header id="masthead" class="site-header" role="banner">
-
 		<nav class="light-blue lighten-1" role="navigation">
     	 <div class="container">
-     		<div class="nav-wrapper"><a id="logo-container" href="<?php echo esc_url( home_url( '/' ) ); ?>" rel="home" class="waves-effect waves-light btn-large white-text"><?php bloginfo( 'name' ); ?><i class="mdi-action-home right"></i></a></a>
-      			<ul id="nav-mobile" class="right side-nav">
+     		<div class="nav-wrapper"><a id="logo-container" href="<?php echo esc_url( home_url( '/' ) ); ?>" rel="home" class="waves-effect waves-light btn-large white-text"><i class="mdi-action-home center"></i></a>
+      			<ul id="nav-mobile" class="left side-nav">
          			 <?php wp_nav_menu( array( 'theme_location' => 'primary' ) ); ?>
        			</ul><a href="#" data-activates="nav-mobile" class="button-collapse"><i class="mdi-navigation-menu"></i></a>
     	  	</div>
    		 </div>
  		</nav>
+ 		<a href="<?php echo esc_url( home_url( '/' ) ); ?>" rel="home" class="waves-effect waves-light btn-large white-text"><h4><?php bloginfo( 'name' ); ?></h4></a>
 	</header><!-- #masthead -->
 
-
-
-	<div id="content" class="site-content container">
+	<div id="content" class="site-content container portage2">