+++ index.php
@@ -13,8 +13,25 @@
 
 get_header(); ?>
 
-	<div id="primary" class="content-area">
+<div class="row">   
+
+   <div class="col s12 m12 l3" id="portagelargeonly">
+				
+				<ul class="section">
+        <li>
+            <div class="section-header">
+            </div>
+                    <div class="section-body">
+							<?php dynamic_sidebar( 'sidebar-3' ); ?>
+					</div><!-- #secondary -->
+        </li>
+      </ul>			
+    </div> 
+
+	<div class="content-area col s12 m10 l7">
+
 		<main id="main" class="site-main" role="main">
+	
 
 		<?php if ( have_posts() ) : ?>
 
@@ -38,9 +55,15 @@ get_header(); ?>
 			<?php get_template_part( 'content', 'none' ); ?>
 
 		<?php endif; ?>
-
+	
 		</main><!-- #main -->
 	</div><!-- #primary -->
+
+
+
+</div>
+
+
 
 <?php get_sidebar(); ?>
 <?php get_footer(); ?>